Update : Fix a bug for StatisticCollector.java causes by a change of OpenFlow protocols, with Parameterized JUnit test
